requirements.txt
requests>=2.31.0
dashscope>=1.14.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
